Delete MFC87D3.tmp
This file appears to have been created by DreamWeaver in the edit directory, and was picked up by the GitHub Windows client as belonging to the repository. We'll have to look into finer-grained control over what gets added to the repository, especially if we want to keep an authorization file apart (and thus out of the public eye).