Slicker Google authorization and first hard-coded query to BigQuery
By use of a Google javascript library, authorization is now handled in a
pop-up window without needing to fuss about browser redirects and URL
hash fragments. However, we now have to find a way to ensure that our
javascript is done loading before the Google javascript finishes. This
suggests we need asynchronous module definition (AMD)? We send the
hard-coded sample query to BigQuery and display the entire result set in
a table generated through jQuery.